@@ -16,11 +16,11 @@
     <url>http://maven.apache.org</url>
 
     <properties>
-        <java.version>1.7</java.version>
-        <spring.version>3.2.0.RELEASE</spring.version>
+        <java.version>1.8</java.version>
+        <spring.version>4.1.4.RELEASE</spring.version>
         <cglib.version>2.2.2</cglib.version>
         <quartz.version>2.1.7</quartz.version>
-        <atomfeed.version>1.9.3</atomfeed.version>
+        <atomfeed.version>1.9.4-SNAPSHOT</atomfeed.version>
         <!--<db.server>10.4.33.112</db.server>-->
         <db.server>localhost</db.server>
         <db.name>openerp</db.name>
@@ -29,6 +29,11 @@
         <db.password>postgres</db.password>
         <db.schema>public</db.schema>
 
+        <junit.version>4.12</junit.version>
+        <mockito-core.version>1.10.19</mockito-core.version>
+        <jackson.version>2.8.1</jackson.version>
+        <httpComponent.version>4.2.5</httpComponent.version>
+        <servlet-api.version>3.0.1</servlet-api.version>
     </properties>
 
     <dependencyManagement>
@@ -145,10 +150,11 @@
                 <artifactId>velocity</artifactId>
                 <version>1.6.2</version>
             </dependency>
+            <!-- https://mvnrepository.com/artifact/org.apache.httpcomponents/httpclient -->
             <dependency>
-                <groupId>commons-httpclient</groupId>
-                <artifactId>commons-httpclient</artifactId>
-                <version>3.1</version>
+                <groupId>org.apache.httpcomponents</groupId>
+                <artifactId>httpclient</artifactId>
+                <version>${httpComponent.version}</version>
             </dependency>
             <dependency>
                 <groupId>log4j</groupId>
@@ -158,17 +164,26 @@
             <dependency>
                 <groupId>javax.servlet</groupId>
                 <artifactId>servlet-api</artifactId>
-                <version>3.0-alpha-1</version>
+                <version>${servlet-api.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.codehaus.jackson</groupId>
-                <artifactId>jackson-core-asl</artifactId>
-                <version>1.5.0</version>
+                <groupId>com.fasterxml.jackson.core</groupId>
+                <artifactId>jackson-databind</artifactId>
+                <version>${jackson.version}</version>
+            </dependency>
+
+            <!--test libraries-->
+            <dependency>
+                <groupId>junit</groupId>
+                <artifactId>junit</artifactId>
+                <version>${junit.version}</version>
+                <scope>test</scope>
             </dependency>
             <dependency>
-                <groupId>org.codehaus.jackson</groupId>
-                <artifactId>jackson-mapper-asl</artifactId>
-                <version>1.5.0</version>
+                <groupId>org.mockito</groupId>
+                <artifactId>mockito-core</artifactId>
+                <version>${mockito-core.version}</version>
+                <scope>test</scope>
             </dependency>
 
         </dependencies>
