@@ -1,6 +1,7 @@
 package org.bahmni.feed.openerp;
 
-import org.codehaus.jackson.map.ObjectMapper;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 public class ObjectMapperRepository {
     public static ObjectMapper objectMapper = new ObjectMapper();
