@@ -1,7 +1,7 @@
 package org.bahmni.feed.openerp.domain.encounter;
 
 
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class OpenMRSDrugOrder {
