@@ -7,7 +7,7 @@ import org.bahmni.feed.openerp.domain.OpenMRSPersonAttributes;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 public class JsonifyTest {
     private OpenMRSPersonAttributes openMRSPersonAttributes;
@@ -31,13 +31,19 @@ public class JsonifyTest {
 
     @Test
     public void testGetJson() throws Exception {
-        assertEquals("{\"attribute1\":\"value1\",\"attribute2\":\"value2\",\"attribute3\":\"value3\"}",
-                openMRSPersonAttributes.toJsonString());
-        assertEquals("{\"address1\":\"address1\",\"address2\":\"address2\",\"address3\":\"address3\"," +
-                        "\"cityVillage\":\"cityVillage\",\"countyDistrict\":\"countyDistrict\"," +
-                        "\"stateProvince\":\"stateProvince\",\"country\":\"country\"}",
-                openMRSPersonAddress.toJsonString());
+        String attributesJsonString = openMRSPersonAttributes.toJsonString();
+        assertTrue(attributesJsonString.contains("\"attribute1\":\"value1\""));
+        assertTrue(attributesJsonString.contains("\"attribute2\":\"value2\""));
+        assertTrue(attributesJsonString.contains("\"attribute3\":\"value3\""));
 
+        String openmrsAddressJsonString = openMRSPersonAddress.toJsonString();
+        assertTrue(openmrsAddressJsonString.contains("\"address1\":\"address1\""));
+        assertTrue(openmrsAddressJsonString.contains("\"address2\":\"address2\""));
+        assertTrue(openmrsAddressJsonString.contains("\"address3\":\"address3\""));
+        assertTrue(openmrsAddressJsonString.contains("\"cityVillage\":\"cityVillage\""));
+        assertTrue(openmrsAddressJsonString.contains("\"countyDistrict\":\"countyDistrict\""));
+        assertTrue(openmrsAddressJsonString.contains("\"stateProvince\":\"stateProvince\""));
+        assertTrue(openmrsAddressJsonString.contains("\"country\":\"country\""));
     }
 
 }
